Remove unused functions from test/utils/manager.go
